requirements.txt
pandas==2.2.0
tabula-py=2.9.0